fix(secu): Authenticated Remote Code Execution in getStats.php
Due to the way preg_replace handles backslahes, the expression:

    function escape_command($command)
    {
        return preg_replace("/[\\\$|`]/", "", $command);
    }

wouldn't work as expected because 3 backslashes are needed to properly refer to the backslash inside the regex.

Therefore the following GET parameter would lead to arbitrary remote code execution:
http://x.x.x.x/centreon/include/Administration/corePerformance/getStats.php?key=active_service_check&ns_id=1=/var/lib/centreon/nagios-perf/perfmon-1/nagios_active_service_execution.rrd:Min:AVERAGE \%26 touch /tmp/das \%26

Because it would be interpreted as \&#039; which still makes it valid for a shell to execute.

I'm not very confident in the multiple layers of filtering in this PHP file. It really needs to NOT use GET parameters.