fix centreontrapd.sdb export on poller when behind RS
when you have a poller linked to a remote server, if you export the trap configuration on it, it will fail because the central server doesn't create the trap directory related to the poller on the remote server. 

this could have been handled using rsync instead of scp (rsync allows you to create parent directories when sending file. Since we use scp, we're better copying the whole directory instead of the sdb file.