Fix LDAP sync deleting users from groups, closes #7142
When group membership is defined at group level, there were two issues with
the syncWithLdap() function that is called periodically, specifically
within listUserForGroup():
* groups were searched in the the user base DN
* the group member attribute was not respected (a hardcoded value of
  "member" was used)

This patch addresses both issues.

The code that is used when a user first logs in makes use of the proper
base DN and group member attribute. Because of this, users were added to the
right groups on their first login, only to be later removed from them.