Update to the SNMP traps documentation
I've updated some of the contents of the SNMP trap documentation. Some of the current documentation is just outdated and some was rewritten for a better understanding (in my opinion).
Most of these updates come from my own difficulty to follow the documentation and implement management by SNMP traps without the help from someone else.