fix(version): Allows to compare API versions using the comma separator of a floating number. (#9075)
Allows to compare API versions using the comma separator of a floating number.
Fixes # (issue)
The problem was converting strings to a floating number. In US the numeric separator is a point but in French it is a comma. The comparison method tried to compare floating numbers with '.' and ',' at the same time.