fix conf export fail because of centreontrapd
on a centreon central server that is receiving a lot of snmp traps, the configuration export will fail most of the time because the /var/lib/centreon/centcore.pm file owner is centreon:centreon with a chmod 644 that doesn't let apache write into this file. 

This situation shouldn't happen because centreontrapd must write in `/var/lib/centreon/centcore/<date>-traps` files. But the removed line overwrite the case where we try to write in those files.