fix(curves) Correction of some form errors and wrong characters displayed (#10370)
Fixed the 'Linked host services' field not being set to the saved value when editing.
Fixed checkboxes not being taken into account when deselected.
Fixed the reset button to reset the list of selection fields in the 'know metrics' form.
Removal of '/n' characters at the bottom of the page